backend/requirements.txt
fastapi[all]
uvicorn[standard]
sqlalchemy
asyncpg
alembic
pydantic-settings
python-jose[cryptography]
passlib[argon2]
python-multipart
